organized public folder, added gallery images
