+++ src/data/scrapbook.js
@@ -1,0 +1,31 @@
+import img1 from "@/public/scrapbook/Community-Garden-at-Sunset.webp";
+import img2 from "@/public/scrapbook/Growing-seedlings.webp";
+import img3 from "@/public/scrapbook/IMG-4576.webp";
+import img4 from "@/public/scrapbook/IMG-5037.webp";
+import img5 from "@/public/scrapbook/IMG-5042.webp";
+import img6 from "@/public/scrapbook/Irrigation-and-Planting.webp";
+import img7 from "@/public/scrapbook/Load-of-Fresh-Produce.webp";
+import img8 from "@/public/scrapbook/Planting-Season.webp";
+import img9 from "@/public/scrapbook/Strawberry-Season.webp";
+import img10 from "@/public/scrapbook/UCR-Community-Garden-Art.webp";
+import img11 from "@/public/scrapbook/UCR-Community-Garden.webp";
+import img12 from "@/public/scrapbook/UCR-Community-Plots.webp";
+import img13 from "@/public/scrapbook/Zucchini-Harvest.webp";
+import img14 from "@/public/scrapbook/Zucchini-Plant.webp";
+
+export const Scrapbook = [
+  img1,
+  img2,
+  img3,
+  img4,
+  img5,
+  img6,
+  img7,
+  img8,
+  img9,
+  img10,
+  img11,
+  img12,
+  img13,
+  img14,
+];
